requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect2==0.7.0
Django==5.1.1
exceptiongroup==1.2.2
fastapi==0.114.0
fonttools==4.54.1
frozenlist==1.4.1
greenlet==3.1.0
h11==0.14.0
html5lib==1.1
httptools==0.6.1
idna==3.8
multidict==6.1.0
mysqlclient==2.2.4
pillow==10.4.0
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.16.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
sqlparse==0.5.1
starlette==0.38.5
tinycss2==1.3.0
twilio==9.3.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
watchfiles==0.24.0
weasyprint==62.3
webencodings==0.5.1
websockets==13.0.1
yarl==1.13.1
zopfli==0.2.3
